move cli out of experiemntal
